just needed perl compatible grep mode to fix it

+++ scripts/filtering-gffs.sh
@@ -6,16 +6,19 @@
 
 DIR=$(dirname $1)
 BASE=$(basename $1 .gff)
+echo "Doing $DIR/$BASE.gff"
 
+set -x
 echo get rid of empty lines
-egrep -v '^$' $1 > $DIR/$BASE.temp
+egrep -v '^$' $1 > "$DIR/$BASE.temp"
 echo get rid of comments
-egrep -v '^#' $DIR/$BASE.temp > $DIR/$BASE.temp2
+egrep -v '^#' $DIR/$BASE.temp > "$DIR/$BASE.temp2"
+file "$DIR/$BASE.temp2"
 echo only get CDS
-egrep "\tCDS\t" $DIR/$BASE.temp2 > $DIR/$BASE-CDS.gff
+grep -P "\tCDS\t" "$DIR/$BASE.temp2" > "$DIR/$BASE-CDS.gff"
 echo "get the rRNA for cuffdiff (or other progs) that support filtering by rRNA seq"
-egrep "\trRNA\t" $DIR/$BASE.temp2 > $DIR/$BASE-rRNA.gff 
+grep -P "\trRNA\t" "$DIR/$BASE.temp2" > "$DIR/$BASE-rRNA.gff" 
 
-echo removing temp files
-rm $DIR/$BASE.temp
-rm $DIR/$BASE.temp2
+echo removing temp files if everything went sucessfull
+[[ -s $DIR/$BASE-CDS.gff ]] && rm $DIR/$BASE.temp
+[[ -s $DIR/$BASE-CDS.gff ]] && rm $DIR/$BASE.temp2
